# src/LupiraHealthApi/Endpoints/MeEndpoints.cs
using LupiraHealthApi.Core.Dtos.Me;
using LupiraHealthApi.Core.Dtos.Records;
using LupiraHealthApi.Handlers;

namespace LupiraHealthApi.Endpoints;

public static class MeEndpoints
{
    public static IEndpointRouteBuilder MapMe(this IEndpointRouteBuilder app)
    {
        app.MapGet("/me", (MeHandler h, CancellationToken ct) => h.GetAsync(ct))
            .RequireAuthorization("ApiPolicy").WithTags("Me")
            .WithSummary("The caller's resolved local identity (JIT-provisioned on first login).")
            .Produces<MeDto>(StatusCodes.Status200OK).Produces(StatusCodes.Status401Unauthorized)
            .WithName("WhoAmI");

        app.MapPost("/me/bootstrap", (MeHandler h, CancellationToken ct) => h.BootstrapAsync(ct))
            .RequireAuthorization("ApiPolicy").WithTags("Me")
            .WithSummary("Idempotently ensure the caller has a personal health record; returns it.")
            .Produces<HealthRecordDto>(StatusCodes.Status200OK).Produces(StatusCodes.Status401Unauthorized)
            .WithName("BootstrapMe");
        return app;
    }
}